requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
argcomplete==3.2.2
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
Cerberus==1.3.5
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coverage==7.4.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
empy==3.3.4
faiss-cpu==1.9.0
fastapi==0.115.1
fastapi-cli==0.0.5
frozendict==2.4.5
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
importlib-resources==6.1.1
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
kconfiglib==14.1.0
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.10
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langsmith==0.1.134
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.22.0
mdurl==0.1.2
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
numpy==1.26.4
nunavut==2.3.1
openai==1.51.2
orjson==3.10.7
packaging==24.1
pandas==2.2.3
peewee==3.17.6
pkgconfig==1.5.5
platformdirs==4.3.6
propcache==0.2.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydsdl==1.20.1
Pygments==2.18.0
PyJWT==2.9.0
pymavlink==2.4.41
pymongo==4.10.1
pyros-genmsg==0.5.8
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
pyulog==1.0.2
PyYAML==6.0.2
redis==5.1.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.38.6
tenacity==8.5.0
tiktoken==0.8.0
tqdm==4.66.5
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.20.0
watchfiles==0.24.0
webencodings==0.5.1
websockets==13.1
yarl==1.15.0
yfinance==0.2.44
psycopg2==2.9.9